Removed inaccurate root indicator (busyBox) check
Assuming the intent of this code is to check if the device is rooted it is causing false positives. This update will remove the check for an indicator (BusyBox) that is known and documented (by the RootBeer author) to cause false positives. This update will not remove other checks used to indicate a rooted device.

This update changes the check from `rootBeer.isRootedWithBusyBoxCheck();` to `rootBeer.isRooted();`. 

This will allow OnePlus (and other) devices owners with non-rooted phones to run the app without getting a "The device is rooted."

Issue documented here: https://github.com/terra-money/station-mobile/issues/11